added vorlesungsunterlagen

+++ vorlesungsdatein/termin_01/02_division.c
@@ -10,8 +10,9 @@ int main()
     int zahl1 = 1, zahl2 = 2;
     double erg;
 
-    erg = (double) zahl1 / zahl2;
-
+    erg = zahl1 / zahl2;
+    // Vermutung 1: erg=0.5
+    // Vermutung 2: erg=1
     printf("%d / %d = %lf", zahl1, zahl2, erg);
 
     return 0;
